+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### vNEXT
 
+- Added support for ApolloClient `v0.5.X` ([PR #])
+- Added `subscribeToMore` function ([PR](https://github.com/kamilkisiela/apollo-client-rxjs/pull/5))
+- **BREAKING CHANGE** No no longer support ApolloClient `v0.4.X`
 - **BREAKING CHANGE** Removed `Apollo` decorator (use `Angular2Apollo` service)
 - **BREAKING CHANGE** Removed `ApolloQueryPipe` (use `SelectPipe` instead)